[#11] Rename `neofsid` contract to `frostfsid`
