+++ pkg/morph/client/frostfsid/addrm_keys.go
@@ -1,4 +1,4 @@
-package neofsid
+package frostfsid
 
 import (
 	"fmt"
