+++ pkg/morph/client/frostfsid/keys.go
@@ -1,4 +1,4 @@
-package neofsid
+package frostfsid
 
 import (
 	"crypto/elliptic"
